Minor improvements and bugs solved

+++ algo_collusion_mm/agents/makers/exp3.py
@@ -163,5 +163,5 @@ class MakerEXP3(Maker):
 
     def reset(self) -> None:
         super().reset()
-        self.weights = np.zeros(self.n_arms)
+        self.weights = np.zeros(self.n_arms, dtype=np.float64)
         return
